bigqueryconnection/v1/python/google-cloud-bigquery-connection/requirements.txt
google-cloud-bigquery-connection==1.19.0
